매너온도 프론트 구현 및 프로필 수정 완료

+++ finalkh11/src/main/java/com/kh/finalkh11/dto/MemberDto.java
@@ -13,7 +13,7 @@ public class MemberDto {
 	private String memberLevel;
 	private String memberEmail;
 	private String memberGender;
-	private int memberManner;
+	private double memberManner;
 	private String memberBirth;
 	private long imgNo;
 }
